Add files via upload
This is an updated index file. I redesigned the layout and added a lot of new features, including checks to make sure you set a quest level, a max room count, a last room button for when you are exiting the dungeon, rat traps, skull traps, and randomized where placement begins in a room.